Fix file names of tests
